+++ internal/xslices/clone_1.20.go
@@ -1,0 +1,14 @@
+//go:build !go1.21
+// +build !go1.21
+
+package xslices
+
+// Clone returns a copy of the slice.
+// The elements are copied using assignment, so this is a shallow clone.
+func Clone[S ~[]E, E any](s S) S {
+	// Preserve nil in case it matters.
+	if s == nil {
+		return nil
+	}
+	return append(S([]E{}), s...)
+}