+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-* PreferLocations is pure function now
+* Refactored the `balancers.PreferLocations()' function - it is a clean/pure function
 
 ## v3.74.2
 * Added description to scan errors with use query service client scanner