+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Refactored the `balancers.PreferLocations()' function - it is a clean/pure function
+
 ## v3.74.2
 * Added description to scan errors with use query service client scanner