+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * Removed check the node is available for query and table service sessions 
 * Refactored the `balancers.PreferLocations()` function - it is a clean/pure function
 * Added experimental `balancers.WithNodeID()` context modifier for define per request the YDB endpoint by NodeID
+* Reverted the allowing the casts from signed YDB types to unsigned destination types if source value is not negative
 
 ## v3.74.2
 * Added description to scan errors with use query service client scanner